+++ src/Luban.Job.Common/Source/TypeVisitors/ProtobufTypeNameVisitor.cs
@@ -98,7 +98,7 @@ namespace Luban.Job.Common.TypeVisitors
 
         public string Accept(TDateTime type)
         {
-            return "int32";
+            return "int64";
         }
 
         public string Accept(TBean type)
